Split code generators into several files
